requirements.txt
# Core
python-dotenv==1.1.0
pyyaml==6.0.2
pydantic==2.11.1

# Data
duckdb==1.2.2
pandas==2.2.3
numpy==1.26.4
pyarrow==19.0.1
scipy==1.15.2
tabulate==0.9.0

# LLM
anthropic==0.52.0

# API
fastapi==0.115.12
uvicorn==0.34.2

# UI
streamlit==1.44.1
plotly==6.0.1
altair==5.5.0

# Eval
scikit-learn==1.6.1

# Dev
pytest==8.3.5
pytest-asyncio==0.26.0
